Change Mmr Size type to u128

+++ src/core/common.cairo
@@ -1,2 +1,2 @@
 type MmrId = u128;
-type MmrSize = usize;
+type MmrSize = u128;
